added target-design Doxygen documentation in actions.h
added target-design Doxygen documentation in calculations.h

added target-design Doxygen documentation in context.h

added target-design Doxygen documentation in deparser.h

added target-design Doxygen documentation in data.h

added target-design Doxygen documentation in counters.h

added target-design Doxygen documentation in dev_mgr.h

added target-design Doxygen documentation in event_logger.h

added target-design Doxygen documentation in fields.h and bytecontainer.h

added target-design Doxygen documentation in field_lists.h

added target-design Doxygen documentation in headers.h

added target-design Doxygen documentation in header_stacks.h

added target-design Doxygen documentation in learning.h

added target-design Doxygen documentation in logger.h

added target-design Doxygen documentation in meters.h

added target-design Doxygen documentation in named_p4object.h

added target-design Doxygen documentation in packet.h

added target-design Doxygen documentation in packet_buffer.h

added target-design Doxygen documentation in parser.h

added target-design Doxygen documentation in phv.h

added target-design Doxygen documentation in pipeline.h

added target-design Doxygen documentation simple_pre.h and simple_pre_lag.h

added target-design Doxygen documentation in queue.h

added target-design Doxygen documentation in stateful.h

added target-design Doxygen documentation in switch.h
