+++ modules/bm_sim/include/bm_sim/event_logger.h
@@ -18,6 +18,8 @@
  *
  */
 
+//! @file event_logger.h
+
 #ifndef BM_SIM_INCLUDE_BM_SIM_EVENT_LOGGER_H_
 #define BM_SIM_INCLUDE_BM_SIM_EVENT_LOGGER_H_
 
@@ -48,6 +50,19 @@ class ActionFn;
 
 typedef uint64_t entry_handle_t;
 
+//! Signals significant packet event by publishing a message on a
+//! transport. This is meant to be used with the nanomsg PUB/SUB transport.
+//! Other processes can subscribe to the channel to monitor the activity of the
+//! switch (e.g. for logging). So far, we are mostly using this for the
+//! end-to-end testing of target switch implementations. Note that depending on
+//! the transport, some messages can be lost (the nanomsg PUB/SUB transport does
+//! not guarantee delivery to all subscribers, and message drops will happen if
+//! a subscriber is slower than the producer).
+//!
+//! Most messages are generated by the bmv2 code, but the target is for now
+//! responsible of generated "packet in" and "packet out" messages (when a
+//! packet is received / transmitted). Obviously, this is optional and you do
+//! not have to do it if you are not interested in using the event logger.
 class EventLogger {
  public:
   explicit EventLogger(std::unique_ptr<TransportIface> transport,
@@ -55,7 +70,9 @@ class EventLogger {
       : transport_instance(std::move(transport)), device_id(device_id) { }
 
   // we need the ingress / egress ports, but they are part of the Packet
+  //! Signal that a packet was received by the switch
   void packet_in(const Packet &packet);
+  //! Signal that a packet was transmitted by the switch
   void packet_out(const Packet &packet);
 
   void parser_start(const Packet &packet, const Parser &parser);
@@ -98,6 +115,13 @@ class EventLogger {
 
 }  // namespace bm
 
+//! Log an event with the event logger.
+//! For example:
+//! @code
+//! BMELOG(packet_in, packet);
+//! // packet processing
+//! BMELOG(packet_out, packet);
+//! @endcode
 #ifdef BMELOG_ON
 #define BMELOG(fn, ...) bm::EventLogger::get()->fn(__VA_ARGS__)
 #else
