+++ modules/bm_sim/include/bm_sim/fields.h
@@ -18,6 +18,8 @@
  *
  */
 
+//! @file fields.h
+
 #ifndef BM_SIM_INCLUDE_BM_SIM_FIELDS_H_
 #define BM_SIM_INCLUDE_BM_SIM_FIELDS_H_
 
@@ -32,6 +34,11 @@
 
 namespace bm {
 
+//! Field objects are used to represent P4 fields. Each Field instance belongs
+//! to a Header instance. When defining your own target, you will have to
+//! manipulate Field objects to set and access the target's intrinsic metadata.
+//! Most of the interesting methods for a target designer are inherited from
+//! Data.
 class Field : public Data {
  public:
   // Data() is called automatically
@@ -56,14 +63,21 @@ class Field : public Data {
     DEBUGGER_NOTIFY_UPDATE(*packet_id, my_id, bytes.data(), nbits);
   }
 
+  //! Return the byte representation of this field. Note that this returns a
+  //! reference to a byte container, which is only valid as long as the Field
+  //! instance is alive.
   const ByteContainer &get_bytes() const {
     return bytes;
   }
 
+  //! Get the number of bytes occupied by this field, not based on its layout in
+  //! the packet header, but assuming a byte boundary alignment for the most
+  //! significant bit. `nbytes = (nbits + 7) / 8`
   int get_nbytes() const {
     return nbytes;
   }
 
+  //! Get the number of bits occupied by this field in the packet header
   int get_nbits() const {
     return nbits;
   }
