+++ modules/bm_sim/include/bm_sim/port_monitor.h
@@ -19,6 +19,8 @@
  *
  */
 
+//! @file port_monitor.h
+
 #ifndef BM_SIM_INCLUDE_BM_SIM_PORT_MONITOR_H_
 #define BM_SIM_INCLUDE_BM_SIM_PORT_MONITOR_H_
 
@@ -26,12 +28,16 @@
 
 namespace bm {
 
+//! Used by DevMgr to monitor the ports attached to the switch
 class PortMonitorIface {
  public:
+  //! Representation of a port number
   typedef unsigned int port_t;
 
+  //! Represents the status of a port
   enum class PortStatus { PORT_ADDED, PORT_REMOVED, PORT_UP, PORT_DOWN };
 
+  //! Signature of the cb function to call when a port status changes
   typedef std::function<void(port_t port_num, const PortStatus status)>
       PortStatusCb;
