+++ modules/bm_sim/include/bm_sim/stateful.h
@@ -18,6 +18,20 @@
  *
  */
 
+//! @file stateful.h
+//! Includes both the Register and RegisterArray classes. RegisterArray is
+//! exposed in P4 v1.0.2 as the `register` object. Action primitives can take a
+//! RegisterArray reference as a parameter (but not a Register though). Here is
+//! a possible implementation of the `register_write()` P4 primitive:
+//! @code
+//! class register_write
+//!   : public ActionPrimitive<RegisterArray &, const Data &, const Data &> {
+//!   void operator ()(RegisterArray &dst, const Data &idx, const Data &src) {
+//!     dst[idx.get_uint()].set(src);
+//!   }
+//! };
+//! @endcode
+
 #ifndef BM_SIM_INCLUDE_BM_SIM_STATEFUL_H_
 #define BM_SIM_INCLUDE_BM_SIM_STATEFUL_H_
 
@@ -30,6 +44,8 @@
 
 namespace bm {
 
+//! A Register object is essentially just a Data object, meant to live in a
+//! RegisterArray. Use the Data class methods to read and write to a Register.
 class Register : public Data {
  public:
   enum RegisterErrorCode {
@@ -63,6 +79,17 @@ class Register : public Data {
 
 typedef p4object_id_t register_array_id_t;
 
+//! RegisterArray corresponds to the `register` standard P4 v1.02 object. A
+//! RegisterArray reference can be used as a P4 primitive parameter. For
+//! example:
+//! @code
+//! class register_write
+//!   : public ActionPrimitive<RegisterArray &, const Data &, const Data &> {
+//!   void operator ()(RegisterArray &dst, const Data &idx, const Data &src) {
+//!     dst[idx.get_uint()].set(src);
+//!   }
+//! };
+//! @endcode
 class RegisterArray : public NamedP4Object {
  public:
   typedef std::vector<Register>::iterator iterator;
@@ -76,25 +103,45 @@ class RegisterArray : public NamedP4Object {
       registers.emplace_back(bitwidth);
   }
 
+  //! Access the register at position \p idx, asserts if bad \p idx
   Register &operator[](size_t idx) {
     assert(idx < size());
     return registers[idx];
   }
 
+  //! @copydoc operator[]
   const Register &operator[](size_t idx) const {
     assert(idx < size());
     return registers[idx];
   }
 
+  //! Access the register at position \p idx, throws a std::out_of_range
+  //! exception if \p idx is invalid
+  Register &at(size_t idx) {
+    return registers.at(idx);
+  }
+
+  //! @copydoc at
+  const Register &at(size_t idx) const {
+    return registers.at(idx);
+  }
+
   // iterators
+
+  //! NC
   iterator begin() { return registers.begin(); }
 
+  //! NC
   const_iterator begin() const { return registers.begin(); }
 
+  //! NC
   iterator end() { return registers.end(); }
 
+  //! NC
   const_iterator end() const { return registers.end(); }
 
+  //! Return the size of the RegisterArray (i.e. number of registers it
+  //! includes)
   size_t size() const { return registers.size(); }
 
  private:
